% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{find_ybr_tiffs}
\alias{find_ybr_tiffs}
\title{Find all the Yoshi behaviour tiffs in a directory hierarchy}
\usage{
find_ybr_tiffs(path = ".", recursive = TRUE, full.names = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}
}
\value{
A data.frame containing information about the identified tiffs
}
\description{
Find all the Yoshi behaviour tiffs in a directory hierarchy
}
\examples{
\dontrun{
tiffdf=find_ybr_tiffs("/path/to/my/flydata")
for(i in 1:nrow(tiffdf)){
  intiff=tiffdf[i,'tiff']
  outpdf=paste0(basename(dirname(intiff)),'.pdf')
  pdf(outpdf)
  plot_smoothed_displacement(intiff)
  dev.off()
  cat(".")
}
}
}
\seealso{
Other read_ybr: \code{\link{read_ybr_summary}};
  \code{\link{read_ybr_tiff}}; \code{\link{read_ybr_xy}}
}

