% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flywatch-package.R
\docType{package}
\name{flywatch-package}
\alias{flywatch}
\alias{flywatch-package}
\title{Behaviour Analysis for Fly Tracking Data}
\description{
For the time being this is focussed on Yoshi Aso's four quadrant
    optogenenetic behaviour devloped at HHMI Janelia Research Campus.
}
\section{Yoshi Behaviour Rig}{


  For the time being this package focuses on analsyis of data derived from
  the four quadrant behaviour rig originally described in
  \href{http://dx.doi.org/10.7554/eLife.04580}{Aso et al. 2014}.

  See the following functions: \itemize{

  \item \code{\link{read_ybr_summary}}

  \item \code{\link{read_ybr_xy}}

  \item \code{\link{plot_ybr}}

  \item \code{\link{plot_smoothed_displacement}}

  }

  for some interesting entry points to this code.
}

\references{
\emph{Mushroom body output neurons encode valence and guide
  memory-based action selection in Drosophila.} Aso Y, Sitaraman D, Ichinose
  T, Kaun KR, Vogt K, Belliart-Guerin G, Placais PY, Robie AA, Yamagata N,
  Schnaitmann C, Rowell WJ, Johnston RM, Ngo TT, Chen N, Korff W, Nitabach
  MN, Heberlein U, Preat T, Branson KM, Tanimoto H, Rubin GM.. \bold{eLIFE}.
  2014 Dec 23;3:e04580. \href{http://dx.doi.org/10.7554/eLife.04580}{doi:
  10.7554/eLife.04580}.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/jefferislab/flywatch/issues}
}

}
\keyword{internal}
