% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{plot_ybr}
\alias{plot_ybr}
\title{Plot time series data derived from Yoshi Behaviour Rig}
\usage{
plot_ybr(x, filter = FALSE, sides = 1, randts = FALSE, lights = c(on1 =
  30, off1 = 60, on2 = 90, off2 = 120), lightcol = rgb(1, 0, 0, alpha = 0.3),
  ...)
}
\arguments{
\item{x}{A (multi-)time series object containing one or more behavioural
variables.}

\item{filter}{\emph{Either} the width in seconds of a simple smoothing filter
\emph{or} a filter defined according to \code{\link[stats]{filter}}. The
signalling value of \code{filter=F} suppresses any filtering.}

\item{sides}{Whether the filter is causal i.e. for past values only (the
default) or centered around lag=0. See \code{\link[stats]{filter}} for
details.}

\item{randts}{Whether to include a randomised version of the time series as a
second plot. Default to true if data are being filtered}

\item{lights}{A length 2 or more vector defining the lights on/off times for
the experiment (in seconds).}

\item{lightcol}{The colour to use to plot the lights on epochs}

\item{...}{Additional arguments to \code{\link[stats]{plot.ts}}}
}
\description{
Plot time series data derived from Yoshi Behaviour Rig
}
\details{
This plot includes an optional smoothed randomised time series which
  is made by randomly permuting the timepoints in the main time series and
  then applying the same smoothing filter. If you want this time series to be
  reliably the same pseudorandom series for publication purposes you may need
  to see the random seed (see \code{\link[base]{set.seed}})
}
\examples{
tiffdf=find_ybr_tiffs(system.file("ybr_tiffs", package='flywatch'))
summdata=read_ybr_summary(tiffdf$tiff[1])
plot_ybr(summdata[,'PIpixel'], filter=FALSE)
plot_ybr(summdata[,'PIpixel'], filter=1)
plot_ybr(summdata[,'PIpixel'], filter=1, randts=TRUE)
plot_ybr(summdata[,c('PIpixel','PIn')], filter=1)
plot_ybr(summdata[,c('PIpixel','PIn')], filter=1)
plot_ybr(summdata[,c('PIpixel', 'PIn')], filter=1, plot.type = "single",
  col=c('red','blue'))
}
\seealso{
\code{\link[stats]{filter}}, \code{\link[stats]{plot.ts}}

Other ybr-plot: \code{\link{plot_smoothed_displacement}}
}

