% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caltech.R
\name{read_caltech_feat}
\alias{read_caltech_feat}
\alias{read_caltech_track}
\title{Read tracking and feature matrices from the Caltech Fly Tracker}
\usage{
read_caltech_feat(f)

read_caltech_track(f)
}
\arguments{
\item{f}{Path to a matlab .mat file}
}
\value{
For \code{read_caltech_feat} an array with dimensions time x features
  x flies

For \code{read_caltech_track} an array with dimensions n_frames x
  n_fields x n_flies and attributes:

  \itemize{

  \item \code{track.mat} raw tracking data (e.g. position, orientation, left
  wing angle)

  \item \code{trk.names} 1 x n_fields cell names of fields in trk.data

  \item \code{trk.flags} n_flags x 6 potential identity swaps (fly1 fly2
  start_fr #' end_fr ambig) The 6th column is mysterious.

  }
}
\description{
Read tracking and feature matrices from the Caltech Fly Tracker
}
\examples{
\dontrun{
feats=read_caltech_feat('~/projects/Dana/tracking/samples/210517_006-feat.mat')
plot(ts(feats[,"vel", "fly001"], start=0, frequency=30))
}
\dontrun{
track=read_caltech_track('~/projects/Dana/tracking/samples/210517_006-track.mat')
plot(ts(track[,c("pos x", "pos y"), "fly001"], start=0, frequency=30))

# plot 2 flies
flies=attr(track, 'flies_in_chamber')
plot(track[,c("pos x", "pos y"), flies[[1]][1]], type='l', col='red', asp=1)
lines(track[,c("pos x", "pos y"), flies[[1]][2]], type='l', col='blue')
}
}
\seealso{
\url{http://www.vision.caltech.edu/Tools/FlyTracker/documentation.html}
}
