% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{read_ybr_summary}
\alias{read_ybr_summary}
\title{Read the summary values from yoshi behaviour TIFF}
\usage{
read_ybr_summary(f, start = 0, frequency = 30, ...)
}
\arguments{
\item{f}{Path to a TIFF file on disk \bold{or} the data from a TIFF that has
previously been read into R}

\item{start}{Time of the first observation}

\item{frequency}{Frequency (i.e. frame rate) for the behavioural observations}

\item{...}{Additional arguments passed to \code{\link[stats]{ts}}.}
}
\value{
a multi timeseries (\code{\link[stats]{ts}}) object containing 26
  columns reporting frame by frame summary statistics.
}
\description{
Read the summary values from yoshi behaviour TIFF
}
\examples{
\dontrun{
some_summ=read_ybr_summary("some.tif")
plot(some_summ[,c(1,2,7)])
}
}
\seealso{
\code{\link[stats]{ts}}

Other read_ybr: \code{\link{read_ybr_tiff}};
  \code{\link{read_ybr_xy}}
}

