% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{read_ybr_xy}
\alias{read_ybr_xy}
\title{Read xy positions from a single behaviour summary file}
\usage{
read_ybr_xy(f, ...)
}
\arguments{
\item{f}{Path to a TIFF file on disk \bold{or} the data from a TIFF that has
previously been read into R}

\item{...}{Additional arguments passed to \code{\link{read_ybr_tiff}} or
eventually to \code{\link[tiff]{readTIFF}}.}
}
\description{
Read xy positions from a single behaviour summary file
}
\seealso{
Other read_ybr: \code{\link{read_ybr_summary}};
  \code{\link{read_ybr_tiff}}
}

