% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{ybr_median_displacement}
\alias{ybr_median_displacement}
\title{Median displacement between nearest neighbour flies in sequential frames}
\usage{
ybr_median_displacement(xy, start = 0, frequency = 30)
}
\arguments{
\item{xy}{List of xy locations per frame produced by
\code{\link{read_ybr_xy}} \emph{or} raw TIFF data or path to TIFF file.}

\item{start}{Time of the first observation}

\item{frequency}{Frequency (i.e. frame rate) for the behavioural observations}
}
\value{
time series
}
\description{
Median displacement between nearest neighbour flies in sequential frames
}
\seealso{
Other ybr.displacement: \code{\link{plot_smoothed_displacement}};
  \code{\link{ybr_raw_displacements}}
}

