% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{ybr_raw_displacements}
\alias{ybr_raw_displacements}
\title{Raw displacements between nearest neighbour flies in sequential frames}
\usage{
ybr_raw_displacements(xy)
}
\arguments{
\item{xy}{List of xy locations per frame produced by
\code{\link{read_ybr_xy}} \emph{or} raw TIFF data or path to TIFF file.}
}
\description{
Each frame may have a different number of "flies". So all we do is report for
each frame \code{n}  the nearest neighbour distance to objects in frame
\code{n+1}. For each frame there will therefore be as many distances are
there are recognised flies.
}
\seealso{
Other ybr.displacement: \code{\link{plot_smoothed_displacement}};
  \code{\link{ybr_median_displacement}}
}

