% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{plot_smoothed_displacement}
\alias{plot_smoothed_displacement}
\title{Plot smoothed displacement estimate}
\usage{
plot_smoothed_displacement(xy, filter = 1, ...)
}
\arguments{
\item{xy}{List of xy locations per frame produced by
\code{\link{read_ybr_xy}} \emph{or} raw TIFF data or path to TIFF file.}

\item{filter}{\emph{Either} the width in seconds of a simple smoothing filter
\emph{or} a filter defined according to \code{\link[stats]{filter}}. The
signalling value of \code{filter=F} suppresses any filtering. Defaults to
1s boxcar filter.}

\item{...}{Additional arguments to \code{\link[stats]{plot.ts}}}
}
\description{
Plot smoothed displacement estimate
}
\examples{
tiffdf=find_ybr_tiffs(system.file("ybr_tiffs", package='flywatch'))
plot_smoothed_displacement(tiffdf$tiff[1])
}
\seealso{
\code{\link[stats]{filter}}, \code{\link[stats]{plot.ts}}

Other ybr-displacement: \code{\link{ybr_median_displacement}};
  \code{\link{ybr_raw_displacements}}

Other ybr-plot: \code{\link{plot_ybr}}
}

