% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{plot_smoothed_displacement}
\alias{plot_smoothed_displacement}
\title{Plot smoothed displacement estimate}
\usage{
plot_smoothed_displacement(xy, filter = 1, lights = c(on1 = 30, off1 = 60,
  on2 = 90, off2 = 120), lightcol = rgb(1, 0, 0, alpha = 0.3), ...)
}
\arguments{
\item{xy}{List of xy locations per frame produced by
\code{\link{read_ybr_xy}} \emph{or} raw TIFF data or path to TIFF file.}

\item{filter}{\emph{Either} the width in seconds of a simple smoothing filter
\emph{or} a filter defined according to \code{\link[stats]{filter}}.}

\item{lights}{A length 2 or more vector defining the lights on/off times for
the experiment (in seconds).}

\item{lightcol}{The colour to use to plot the lights on epochs}

\item{...}{Additional arguments to \code{plot}}
}
\description{
Plot smoothed displacement estimate
}
\seealso{
\code{\link[stats]{filter}}

Other ybr-displacement: \code{\link{ybr_median_displacement}};
  \code{\link{ybr_raw_displacements}}
}

