% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caltech.R
\name{read_caltech_feat}
\alias{read_caltech_feat}
\title{Read the feature matrix from the Caltech Fly Tracker}
\usage{
read_caltech_feat(f)
}
\arguments{
\item{f}{Path to a matlab .mat file}
}
\value{
an array with dimensions time x features x flies
}
\description{
Read the feature matrix from the Caltech Fly Tracker
}
\examples{
\dontrun{
feats=read_caltech_feat('~/projects/Dana/tracking/samples/210517_006-feat.mat')
plot(ts(feats[,"vel", "fly001"], start=0, frequency=30))
}
}
\seealso{
\url{http://www.vision.caltech.edu/Tools/FlyTracker/documentation.html}
}
