% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yoshi_behav_functions.R
\name{read_ybr_tiff}
\alias{read_ybr_tiff}
\title{Read in one of Yoshi's behaviour TIFF summary files}
\usage{
read_ybr_tiff(f, all = TRUE, as.is = FALSE, ...)
}
\arguments{
\item{f}{Path to a TIFF file on disk \bold{or} the data from a TIFF that has
previously been read into R.}

\item{all}{TIFF files can contain more than one image, if \code{all =
    TRUE} then all images are returenin a list, otherwsie only the first
  image is returned.}

\item{as.is}{attempt to return original values without re-scaling
  where possible}

\item{...}{Additional arguments passed to \code{\link[tiff]{readTIFF}}}
}
\value{
list of TIFF slices with extra class \code{ybr_raw}
}
\description{
Low level function to read in raw data from a Yoshi behaviour
  summary TIFF file. You can use this to customise how the TIFF file is read
  in \emph{or} if you want to extract multiple kinds of information from the
  same TIFF file using different functions.
}
\seealso{
\code{\link[tiff]{readTIFF}}

Other read_ybr: \code{\link{read_ybr_summary}};
  \code{\link{read_ybr_xy}}
}

